only notify ckpt when there is no onging ckpt
